Fixed warnings
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@15131 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ desktop/geogebra/cas/view/LaTeXPanel.java
@@ -1,6 +1,5 @@
 package geogebra.cas.view;
 
-import geogebra.euclidian.DrawEquation;
 import geogebra.main.Application;
 
 import java.awt.Dimension;
@@ -59,7 +58,7 @@ public class LaTeXPanel extends JPanel {
 				app.getPlainFontCommon(), false, new geogebra.awt.Color(getForeground()), 
 				new geogebra.awt.Color(getBackground()), true);	
 
-		return new Dimension((int)fd.getWidth(), (int)fd.getHeight());
+		return new Dimension(fd.getWidth(), fd.getHeight());
 	}
 	
 	private boolean ensureImageSize(int width, int height) {
